refactor(slack-bridge): extract thread confirmation policy
